requirements.txt
slack_sdk
slack_bolt
Flask>=1.1
gunicorn>=20
psycopg2-binary
